+++ react-app/src/assets/Cart Icon.png
