+++ react-app/src/assets/Search Icon.png
