+++ react-app/src/assets/btn-right-arrow.png
