+++ react-app/src/assets/cutOnion.png
