+++ react-app/src/assets/dark-right-arrow.png
