+++ react-app/src/components/image-cards.jsx
@@ -2,7 +2,7 @@ import "../components/styles/image-cards.css";
 export default function ImgCards(props) {
     return (
         <>
-            <article className="img-card" style={{backgroundImage: "../assests/home-section-two-cardImg-one.jpg"}}>
+            <article className="img-card" style={{backgroundImage: "../assets/home-section-two-cardImg-one.jpg"}}>
                 <article className="content">
                     <p className="curvy-text white-text">
                         {props.curvyText}
