+++ react-app/src/components/styles/title-image.css
@@ -5,20 +5,20 @@
     overflow: hidden;
   }
   
-  .title-image-bg {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-  }
-  
-  .title-text {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    font-size: 70px;
-    font-family: 'Roboto';
-    font-weight: 800;
-    color: var(--title-color);
-  }
+.title-image-bg {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.title-image .title-text {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  font-size: 70px;
+  font-family: 'Roboto';
+  font-weight: 800;
+  color: var(--title-color);
+}
