+++ react-app/src/pages/about.jsx
@@ -1,19 +1,19 @@
-import titlebg from '../assests/aboutbg.png';
+import titlebg from '../assets/aboutbg.png';
 import TitleImage from '../components/title-image';
-import about1 from '../assests/about1.png';
-import about2 from '../assests/about2.png';
+import about1 from '../assets/about1.png';
+import about2 from '../assets/about2.png';
 import '../styles/about.css';
 import IconText from '../components/icon-text';
 import PrimaryColoredBtn from '../components/primary-colored-btn';
 
-import icon1 from '../assests/about-icon1.svg';
-import icon2 from '../assests/about-icon2.svg';
-import icon3 from '../assests/about-icon3.svg';
-import icon4 from '../assests/about-icon4.svg';
-import icon5 from '../assests/about-icon5.svg';
-import icon6 from '../assests/about-icon6.svg';
-import icon7 from '../assests/about-icon7.svg';
-import profile1 from '../assests/profile1.png';
+import icon1 from '../assets/about-icon1.svg';
+import icon2 from '../assets/about-icon2.svg';
+import icon3 from '../assets/about-icon3.svg';
+import icon4 from '../assets/about-icon4.svg';
+import icon5 from '../assets/about-icon5.svg';
+import icon6 from '../assets/about-icon6.svg';
+import icon7 from '../assets/about-icon7.svg';
+import profile1 from '../assets/profile1.png';
 import IconText2 from '../components/icontext2';
 import IconText3 from '../components/icontext3';
 import Subscribe from '../components/subscribe';
